+++ adminimize/inc-options/dashboard_options.php
@@ -23,7 +23,7 @@ if ( ! function_exists( 'add_action' ) ) {
 			// get widgets
 			$widgets = _mw_adminimize_get_option_value( 'mw_adminimize_dashboard_widgets' );
 			if ( NULL === $widgets ) {
-				echo '<p class="form-invalid">';
+				echo '<p>';
 				esc_attr_e(
 					'To complete the installation for Dashboard Widgets you must visit your dashboard once and then come back to Settings > Adminimize to configure who has access to each widget.',
 					'adminimize'
@@ -33,6 +33,15 @@ if ( ! function_exists( 'add_action' ) ) {
 				?>
 
 				<table summary="config_edit_dashboard" class="widefat">
+					<colgroup>
+						<?php
+						$col = 0;
+						foreach ( $user_roles_names as $role_name ) {
+							echo '<col class="col' . $col . '">' . "\n";
+							$col ++;
+						}
+						?>
+					</colgroup>
 					<thead>
 					<tr>
 						<th><?php esc_attr_e( 'Option', 'adminimize' ); ?></th>
@@ -45,11 +54,9 @@ if ( ! function_exists( 'add_action' ) ) {
 					<tr>
 						<td><?php esc_attr_e( 'Select all', 'adminimize' ); ?></td>
 						<?php
-						foreach ( $user_roles_names as $role_name ) {
-							$role_name = strtolower( $role_name );
-							$role_name = preg_replace( '/[^a-z0-9]+/', '', $role_name );
+						foreach ( $user_roles as $role_slug ) {
 							echo '<td class="num">';
-							echo '<input id="select_all" class="dashboard_options_' . $role_name
+							echo '<input id="select_all" class="dashboard_options_' . $role_slug
 								. '" type="checkbox" name="" value="" />';
 							echo '</td>' . "\n";
 						} ?>