+++ adminimize/inc-options/write_post_options.php
@@ -20,6 +20,15 @@ if ( ! function_exists( 'add_action' ) ) {
 			<br class="clear" />
 
 			<table summary="config_edit_post" class="widefat">
+				<colgroup>
+					<?php
+					$col = 0;
+					foreach ( $user_roles_names as $role_name ) {
+						echo '<col class="col' . $col . '">' . "\n";
+						$col ++;
+					}
+					?>
+				</colgroup>
 				<thead>
 				<tr>
 					<th><?php esc_attr_e( 'Write options - Post', 'adminimize' ); ?></th>
@@ -32,11 +41,9 @@ if ( ! function_exists( 'add_action' ) ) {
 				<tr>
 					<td><?php esc_attr_e( 'Select all', 'adminimize' ); ?></td>
 					<?php
-					foreach ( $user_roles_names as $role_name ) {
-						$role_name = strtolower( $role_name );
-						$role_name = preg_replace( '/[^a-z0-9]+/', '', $role_name );
+					foreach ( $user_roles as $role_slug ) {
 						echo '<td class="num">';
-						echo '<input id="select_all" class="write_post_options_' . $role_name
+						echo '<input id="select_all" class="write_post_options_' . $role_slug
 							. '" type="checkbox" name="" value="" />';
 						echo '</td>' . "\n";
 					} ?>
@@ -72,7 +79,6 @@ if ( ! function_exists( 'add_action' ) ) {
 						&& 'excerpt' === $post_type_support
 					) {
 						$post_type_support = $post_type . 'excerpt';
-
 					}
 					if ( 'page-attributes' === $post_type_support ) {
 						$post_type_support = 'pageparentdiv';
@@ -192,7 +198,7 @@ if ( ! function_exists( 'add_action' ) ) {
 						foreach ( $user_roles as $role ) {
 							$checked_user_role_[ $role ] = ( isset( $disabled_metaboxes_post_[ $role ] )
 								&& in_array(
-									$metabox, $disabled_metaboxes_post_[ $role ]
+									$metabox, $disabled_metaboxes_post_[ $role ], FALSE
 								) ) ? ' checked="checked"' : '';
 						}
 						echo '<tr>' . "\n";